+++ gulpfile.js
@@ -10,20 +10,23 @@ const gulp = require('gulp'),
 	markdown = require('gulp-markdown'),
 	browserSync = require('browser-sync').create();
 
+const app_base = './app';
 const paths = {
-	styles: {
+	'styles': {
+		// If you have a main SCSS file that imports all your other SCSS files, or else set to empty string
+		'main': app_base + '/scss/main.scss',
 		// By using styles/**/*.sass we're telling gulp to check all folders for any sass file
-		src: 'app/scss/*.scss',
+		'src': app_base + '/scss/*.scss',
 		// Compiled files will end up in whichever folder it's found in (partials are not compiled)
-		dest: 'app/css'
+		'dest': app_base + '/css'
 	},
-	scripts: {
-		src: 'app/js/*.js',
-		dest: 'app/js/min'
+	'scripts': {
+		'src': app_base + '/js/*.js',
+		'dest': app_base + '/js/min'
 	},
-	markdown: {
-		src: './*.md',
-		dest: './'
+	'markdown': {
+		'src': './*.md',
+		'dest': './'
 	}
 };
 const lib_path = './app/js/lib/'; // path to vendor libraries
@@ -36,9 +39,10 @@ const js_src_array = jslib_files.concat(paths.scripts.src); // concats lib files
  * Process SCSS files
  ===================================== */
 function style() {
+	const src = paths.styles.main ? paths.styles.main : paths.styles.src;
 	return (
 		gulp
-			.src(paths.styles.src)
+			.src(src)
 			// Initialize sourcemaps before compilation starts
 			.pipe(sourcemaps.init())
 			.pipe(sass())
@@ -93,8 +97,8 @@ function reload() {
 function watch() {
 	browserSync.init({
 		// You can tell browserSync to use this directory and serve it as a mini-server
-		server: {
-			baseDir: './app'
+		'server': {
+			'baseDir': './app'
 		}
 		// If you are already serving your website locally using something like apache
 		// You can use the proxy setting to proxy that instead
@@ -106,7 +110,7 @@ function watch() {
 	// We should tell gulp which files to watch to trigger the reload
 	// This can be html or whatever you're using to develop your website
 	// Note -- you can obviously add the path to the Paths object
-	//gulp.watch("src/*.html", reload);
+	// gulp.watch("src/*.html", reload);
 	gulp.watch('app/*.html').on('change', browserSync.reload);
 }
 
@@ -131,7 +135,7 @@ var build = gulp.parallel(markdownWatch, style, script, watch);
 /*
  * You can still use `gulp.task` to expose tasks
  */
-//gulp.task('build', build);
+// gulp.task('build', build);
 
 /*
  * Define default task that can be called by just running `gulp` from cli